+++ TeamCode/build.gradle
@@ -14,13 +14,19 @@
 // Include common definitions from above.
 apply from: '../build.common.gradle'
 apply from: '../build.dependencies.gradle'
-apply plugin: 'org.team11260.fast-load-plugin'
+apply plugin: 'dev.frozenmilk.sinister.sloth.Load'
 
 android {
     namespace = 'org.firstinspires.ftc.teamcode'
 
     packagingOptions {
         jniLibs.useLegacyPackaging true
+    }
+
+    // Add these compiler options
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
     }
 }
 
@@ -30,30 +36,63 @@ buildscript {
         maven {
             url = 'https://www.matthewo.tech/maven/'
         }
+        maven {
+            url = 'https://repo.dairy.foundation/snapshots'
+        }
+
     }
-
     dependencies {
-        classpath 'org.team11260:fast-load-plugin:0.1.2'
-
+        classpath "dev.frozenmilk.sinister.sloth:Load:0.0.2"
     }
 }
 
+
 repositories {
+
+    maven {
+        url = 'https://repo.dairy.foundation/releases'
+    }
+    maven {
+        url = 'https://repo.dairy.foundation/snapshots'
+    }
+
+    maven {
+        url = 'https://maven.brott.dev/'
+    }
+
+    maven {
+        url = 'https://jitpack.io'
+    }
     maven {
         url = 'https://www.matthewo.tech/maven/'
     }
+
 }
+
 
 dependencies {
     implementation project(':FtcRobotController')
-    annotationProcessor files('lib/OpModeAnnotationProcessor.jar')
+    implementation('com.acmerobotics.slothboard:dashboard:0.2.1+0.4.16') {
+        exclude group: 'org.firstinspires.ftc'
+    }
 
+    annotationProcessor files('lib/OpModeAnnotationProcessor.jar')
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.12.7'
     implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.4.21'
-
-    implementation 'org.team11260:fast-load:0.1.2'
-
     implementation "org.tensorflow:tensorflow-lite:2.2.0"
     implementation "org.tensorflow:tensorflow-lite-support:0.4.0"
     implementation "org.apache.commons:commons-lang3:3.17.0"
+    implementation 'org.ftclib.ftclib:core:2.1.1'
+
+
+    implementation("dev.frozenmilk.sinister:Sloth:0.2.1") {
+        exclude group: "com.acmerobotics.dashboard"
+    }
+    implementation("dev.frozenmilk.sinister.sloth:DashFix:0.0.0") {
+        exclude group: "com.acmerobotics.dashboard"
+    }
+//    implementation 'dev.frozenmilk.sinister:Sloth:0.2.1'
+//    implementation 'dev.frozenmilk.sinister.sloth:DashFix:0.0.0'
+
+    //implementation("com.acmerobotics.slothboard:dashboard:0.2.1+0.4.16")
 }
