+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Pedro/constants/LConstants.java
@@ -1,4 +1,4 @@
-package pedroPathing.tuners.constants;
+package org.firstinspires.ftc.teamcode.Pedro.constants;
 
 import com.pedropathing.localization.*;
 import com.pedropathing.localization.constants.*;
