+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Pedro/examples/Triangle.java
@@ -1,4 +1,4 @@
-package pedroPathing.tuners.examples;
+package org.firstinspires.ftc.teamcode.Pedro.examples;
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.pedropathing.follower.Follower;
@@ -11,8 +11,8 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 
-import pedroPathing.tuners.constants.FConstants;
-import pedroPathing.tuners.constants.LConstants;
+import org.firstinspires.ftc.teamcode.Pedro.constants.FConstants;
+import org.firstinspires.ftc.teamcode.Pedro.constants.LConstants;
 
 
 /**
