+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Pedro/tuners_tests/localization/MotorDirections.java
@@ -1,4 +1,4 @@
-package pedroPathing.tuners.tuners_tests.localization;
+package org.firstinspires.ftc.teamcode.Pedro.tuners_tests.localization;
 import static com.pedropathing.follower.FollowerConstants.leftFrontMotorDirection;
 import static com.pedropathing.follower.FollowerConstants.leftFrontMotorName;
 import static com.pedropathing.follower.FollowerConstants.leftRearMotorDirection;
@@ -22,8 +22,8 @@ import org.firstinspires.ftc.robotcore.external.Telemetry;
 import java.util.Arrays;
 import java.util.List;
 
-import pedroPathing.tuners.constants.FConstants;
-import pedroPathing.tuners.constants.LConstants;
+import org.firstinspires.ftc.teamcode.Pedro.constants.FConstants;
+import org.firstinspires.ftc.teamcode.Pedro.constants.LConstants;
 
 @TeleOp(name = "Motor Directions", group = "Teleop Test")
 public class MotorDirections extends OpMode {
